+++ backends/xla/xla.go
@@ -78,15 +78,17 @@ import (
 // The stablehlo backend also accepts the "xla", "hlo" and "pjrt" aliases.
 const BackendName = "xla"
 
-// Disable XLA logging by default by setting TF_CPP_MIN_LOG_LEVEL to 2 (errors level), if it is not already set.
+// Disable XLA logging by default by setting TF_CPP_MIN_LOG_LEVEL to 3 (above errors level), if it is not already set.
 // This won't work if the PJRT is linked statically or dynamically before the go program start (without `dlopen` that is).
+//
+// See issue https://github.com/openxla/xla/issues/26466
 func init() {
 	const TensorflowCPPMinLogLevelEnv = "TF_CPP_MIN_LOG_LEVEL"
 	tfLogLevel := os.Getenv(TensorflowCPPMinLogLevelEnv)
 	if tfLogLevel == "" {
-		err := os.Setenv(TensorflowCPPMinLogLevelEnv, "2")
+		err := os.Setenv(TensorflowCPPMinLogLevelEnv, "3")
 		if err != nil {
-			klog.Errorf("Failed to set $%s to 2: %v", TensorflowCPPMinLogLevelEnv, err)
+			klog.Errorf("Failed to set $%s to 3: %v", TensorflowCPPMinLogLevelEnv, err)
 		}
 	}
 }